Separate t function extraction from i18next instances.
